+++ src/components/views/OnlineModList.vue
@@ -64,6 +64,7 @@
 import { Prop, Vue } from 'vue-property-decorator';
 import Component from 'vue-class-component';
 import ThunderstoreMod from '../../model/ThunderstoreMod';
+import GameManager from '../../model/game/GameManager';
 import ManagerSettings from '../../r2mm/manager/ManagerSettings';
 import { ExpandableCard, Link } from '../all';
 import DownloadModModal from './DownloadModModal.vue';
@@ -85,8 +86,7 @@ export default class OnlineModList extends Vue {
     @Prop()
     pagedModList!: ThunderstoreMod[];
 
-    @Prop({required: true})
-    private settings!: ManagerSettings;
+    settings: ManagerSettings = new ManagerSettings();
 
     private cardExpanded: boolean = false;
     private darkTheme: boolean = false;
@@ -137,7 +137,9 @@ export default class OnlineModList extends Vue {
         this.$emit('error', error);
     }
 
-    created() {
+    async created() {
+        this.settings = await ManagerSettings.getSingleton(GameManager.activeGame);
+
         if (this.settingsUpdateTimer !== null) {
             clearInterval(this.settingsUpdateTimer);
         }
